requirements.txt
# ==============================================================================
# VietIDP - Requirements (Local CPU)
# ==============================================================================
# Cài đặt: pip install -r requirements.txt

# --- Core ---
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0
requests>=2.31.0

# --- Document Processing ---
python-docx>=1.0.0

# --- Data & Utils ---
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# --- YOLO Detection ---
ultralytics>=8.0.0

# --- Web API ---
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# --- Visualization ---
matplotlib>=3.8.0
